requirement.txt
amqp==1.4.7
anyjson==0.3.3
Babel==2.0
billiard==3.3.0.20
celery==3.1.18
cx-Oracle==5.2
flower==0.8.1
futures==3.0.2
kombu==3.0.30
meld3==1.0.2
MySQL-python==1.2.5
Naked==0.1.31
pycrypto==2.6.1
pymongo==3.2.1
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.7.0
shellescape==3.4.1
six==1.9.0
sqlparse==0.2.2
supervisor==3.1.3
tornado==4.2.1
wheel==0.24.0
WTForms==2.1
WTForms-JSON==0.2.10